Minor styling tweaks to details panel and landing pages #423
